refactor: extract shared SpanCoversColumn for Signature ops (#955)
Consolidate five identical exclusive-end helpers into SpanCoverage and
delegate TypeSymbolResolver.SpanCoversColumn to the same implementation.
