+++ src/RoslynMcp.Core/Refactoring/Signature/RemoveParameterOperation.cs
@@ -8,6 +8,7 @@ using RoslynMcp.Contracts.Errors;
 using RoslynMcp.Contracts.Models;
 using RoslynMcp.Core.FileSystem;
 using RoslynMcp.Core.Refactoring.Base;
+using RoslynMcp.Core.Resolution;
 using RoslynMcp.Core.Workspace;
 
 namespace RoslynMcp.Core.Refactoring.Signature;
@@ -327,35 +328,11 @@ public sealed class RemoveParameterOperation : RefactoringOperationBase<RemovePa
 
     private static bool MethodCoversColumn(MethodDeclarationSyntax method, int line, int column) =>
         IdentifierCoversColumn(method, line, column) ||
-        SpanCoversColumn(method.GetLocation().GetLineSpan(), line, column);
+        SpanCoverage.SpanCoversColumn(method.GetLocation().GetLineSpan(), line, column);
 
     private static bool IdentifierCoversColumn(MethodDeclarationSyntax method, int line, int column) =>
-        SpanCoversColumn(method.Identifier.GetLocation().GetLineSpan(), line, column);
+        SpanCoverage.SpanCoversColumn(method.Identifier.GetLocation().GetLineSpan(), line, column);
 
-    /// <summary>
-    /// 1-based line/column coverage. <see cref="FileLinePositionSpan.EndLinePosition"/>
-    /// is exclusive, so <paramref name="column"/> must be strictly before the
-    /// exclusive end (reject <c>column &gt;= endCol</c>). Treating the end as
-    /// inclusive would let the first character of an adjacent method also
-    /// match the previous declaration. Same helper as
-    /// <c>ChangeSignatureOperation.SpanCoversColumn</c> /
-    /// <c>AddParameterOperation.SpanCoversColumn</c>.
-    /// </summary>
-    internal static bool SpanCoversColumn(FileLinePositionSpan span, int line, int column)
-    {
-        var startLine = span.StartLinePosition.Line + 1;
-        var endLine = span.EndLinePosition.Line + 1;
-        var startCol = span.StartLinePosition.Character + 1;
-        var endCol = span.EndLinePosition.Character + 1;
-
-        if (line < startLine || line > endLine)
-            return false;
-        if (line == startLine && column < startCol)
-            return false;
-        if (line == endLine && column >= endCol)
-            return false;
-        return true;
-    }
 
     internal static IParameterSymbol FindParameter(IMethodSymbol method, string name)
     {
