+++ src/RoslynMcp.Core/Resolution/SpanCoverage.cs
@@ -1,0 +1,32 @@
+using Microsoft.CodeAnalysis;
+
+namespace RoslynMcp.Core.Resolution;
+
+/// <summary>
+/// Shared 1-based exclusive-end coverage helpers for <see cref="FileLinePositionSpan"/>.
+/// </summary>
+internal static class SpanCoverage
+{
+    /// <summary>
+    /// 1-based line/column coverage. <see cref="FileLinePositionSpan.EndLinePosition"/>
+    /// is exclusive, so <paramref name="column"/> must be strictly before the
+    /// exclusive end (reject <c>column &gt;= endCol</c>). Treating the end as
+    /// inclusive would let the first character of an adjacent declaration also
+    /// match the previous one.
+    /// </summary>
+    internal static bool SpanCoversColumn(FileLinePositionSpan span, int line, int column)
+    {
+        var startLine = span.StartLinePosition.Line + 1;
+        var endLine = span.EndLinePosition.Line + 1;
+        var startCol = span.StartLinePosition.Character + 1;
+        var endCol = span.EndLinePosition.Character + 1;
+
+        if (line < startLine || line > endLine)
+            return false;
+        if (line == startLine && column < startCol)
+            return false;
+        if (line == endLine && column >= endCol)
+            return false;
+        return true;
+    }
+}
