+++ src/RoslynMcp.Core/Resolution/TypeSymbolResolver.cs
@@ -244,30 +244,10 @@ public sealed class TypeSymbolResolver
         SpanCoversColumn(type.Identifier.GetLocation().GetLineSpan(), line, column);
 
     /// <summary>
-    /// 1-based line/column coverage. <see cref="FileLinePositionSpan.EndLinePosition"/>
-    /// is exclusive, so <paramref name="column"/> must be strictly before the
-    /// exclusive end (reject <c>column &gt;= endCol</c>). Treating the end as
-    /// inclusive would let the first character of an adjacent type also
-    /// match the previous declaration. Same helper as
-    /// <c>UseBaseTypeOperation.SpanCoversColumn</c> /
-    /// <c>EncapsulateFieldOperation.SpanCoversColumn</c> /
-    /// <c>PushMembersDownOperation.SpanCoversColumn</c>.
+    /// 1-based line/column coverage. Delegates to <see cref="SpanCoverage.SpanCoversColumn"/>.
     /// </summary>
-    internal static bool SpanCoversColumn(FileLinePositionSpan span, int line, int column)
-    {
-        var startLine = span.StartLinePosition.Line + 1;
-        var endLine = span.EndLinePosition.Line + 1;
-        var startCol = span.StartLinePosition.Character + 1;
-        var endCol = span.EndLinePosition.Character + 1;
-
-        if (line < startLine || line > endLine)
-            return false;
-        if (line == startLine && column < startCol)
-            return false;
-        if (line == endLine && column >= endCol)
-            return false;
-        return true;
-    }
+    internal static bool SpanCoversColumn(FileLinePositionSpan span, int line, int column) =>
+        SpanCoverage.SpanCoversColumn(span, line, column);
 }
 
 /// <summary>
