+++ tests/RoslynMcp.Core.Tests/Refactoring/Signature/AddParameterOperationTests.cs
@@ -6,6 +6,7 @@ using RoslynMcp.Contracts.Errors;
 using RoslynMcp.Contracts.Models;
 using RoslynMcp.Core.Refactoring;
 using RoslynMcp.Core.Refactoring.Signature;
+using RoslynMcp.Core.Resolution;
 using RoslynMcp.Core.Workspace;
 using Xunit;
 
@@ -883,10 +884,10 @@ public class AddParameterOperationTests
         var startCol = span.StartLinePosition.Character + 1;
         var endCol = span.EndLinePosition.Character + 1;
 
-        Assert.True(AddParameterOperation.SpanCoversColumn(span, line, startCol));
-        Assert.True(AddParameterOperation.SpanCoversColumn(span, line, endCol - 1));
-        Assert.False(AddParameterOperation.SpanCoversColumn(span, line, endCol));
-        Assert.False(AddParameterOperation.SpanCoversColumn(span, line, startCol - 1));
+        Assert.True(SpanCoverage.SpanCoversColumn(span, line, startCol));
+        Assert.True(SpanCoverage.SpanCoversColumn(span, line, endCol - 1));
+        Assert.False(SpanCoverage.SpanCoversColumn(span, line, endCol));
+        Assert.False(SpanCoverage.SpanCoversColumn(span, line, startCol - 1));
     }
 
     [SkippableFact]
