+++ tests/RoslynMcp.Core.Tests/Refactoring/Signature/ChangeSignatureOperationTests.cs
@@ -5,6 +5,7 @@ using RoslynMcp.Contracts.Errors;
 using RoslynMcp.Contracts.Models;
 using RoslynMcp.Core.Refactoring;
 using RoslynMcp.Core.Refactoring.Signature;
+using RoslynMcp.Core.Resolution;
 using RoslynMcp.Core.Workspace;
 using Xunit;
 
@@ -428,10 +429,10 @@ public class ChangeSignatureOperationTests
         var startCol = span.StartLinePosition.Character + 1;
         var endCol = span.EndLinePosition.Character + 1;
 
-        Assert.True(ChangeSignatureOperation.SpanCoversColumn(span, line, startCol));
-        Assert.True(ChangeSignatureOperation.SpanCoversColumn(span, line, endCol - 1));
-        Assert.False(ChangeSignatureOperation.SpanCoversColumn(span, line, endCol));
-        Assert.False(ChangeSignatureOperation.SpanCoversColumn(span, line, startCol - 1));
+        Assert.True(SpanCoverage.SpanCoversColumn(span, line, startCol));
+        Assert.True(SpanCoverage.SpanCoversColumn(span, line, endCol - 1));
+        Assert.False(SpanCoverage.SpanCoversColumn(span, line, endCol));
+        Assert.False(SpanCoverage.SpanCoversColumn(span, line, startCol - 1));
     }
 
     #endregion
