+++ tests/RoslynMcp.Core.Tests/Refactoring/Signature/RemoveParameterOperationTests.cs
@@ -6,6 +6,7 @@ using RoslynMcp.Contracts.Errors;
 using RoslynMcp.Contracts.Models;
 using RoslynMcp.Core.Refactoring;
 using RoslynMcp.Core.Refactoring.Signature;
+using RoslynMcp.Core.Resolution;
 using RoslynMcp.Core.Workspace;
 using Xunit;
 
@@ -731,10 +732,10 @@ public class RemoveParameterOperationTests
         var startCol = span.StartLinePosition.Character + 1;
         var endCol = span.EndLinePosition.Character + 1;
 
-        Assert.True(RemoveParameterOperation.SpanCoversColumn(span, line, startCol));
-        Assert.True(RemoveParameterOperation.SpanCoversColumn(span, line, endCol - 1));
-        Assert.False(RemoveParameterOperation.SpanCoversColumn(span, line, endCol));
-        Assert.False(RemoveParameterOperation.SpanCoversColumn(span, line, startCol - 1));
+        Assert.True(SpanCoverage.SpanCoversColumn(span, line, startCol));
+        Assert.True(SpanCoverage.SpanCoversColumn(span, line, endCol - 1));
+        Assert.False(SpanCoverage.SpanCoversColumn(span, line, endCol));
+        Assert.False(SpanCoverage.SpanCoversColumn(span, line, startCol - 1));
     }
 
     [SkippableFact]
